feat: repository 삭제
