+++ src/signature/signature.controller.ts
@@ -10,17 +10,19 @@ import { HomeSignatureDto } from './dto/home-signature.dto';
 export class SignatureController {
   constructor(private readonly signatureService: SignatureService) {}
 
+  /*
   @Get('/')
   getMySignature(@Body() userId: number) {
     // 임시로 토큰이 아닌 유저 아이디 받도록 구현
     console.log('signature/: 내 시그니처 요청');
     return this.signatureService.homeSignature(userId);
   }
+  */
 
   @Post('/new')
   async createNewSignature(
     @Body() newSignature: CreateSignatureDto,
-  ): Promise<boolean> {
+  ): Promise<number> {
     const result = await this.signatureService.createSignature(newSignature);
     return result;
   }
