+++ src/signature/signature.like.entity.ts
@@ -11,7 +11,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import SignatureEntity from './signature.entity';
+import { SignatureEntity } from './signature.entity';
 import { UserEntity } from 'src/user/user.entity';
 
 @Entity()
