+++ src/signature/signature.module.ts
@@ -3,8 +3,7 @@
 import { Module } from '@nestjs/common';
 import { SignatureService } from './signature.service';
 import { SignatureController } from './signature.controller';
-import { SignatureEntityRepository } from './signature.repository';
-import SignatureEntity from './signature.entity';
+import { SignatureEntity } from './signature.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { DataSource } from 'typeorm';
 import { EntityManager } from 'typeorm';
@@ -12,6 +11,6 @@ import { EntityManager } from 'typeorm';
 @Module({
   //imports: [TypeOrmModule.forFeature([SignatureEntity])],
   controllers: [SignatureController],
-  providers: [SignatureService, SignatureEntityRepository, EntityManager],
+  providers: [SignatureService, EntityManager],
 })
 export class SignatureModule {}
